Gems/AWSMetrics/cdk/requirements.txt
aws-cdk.core>=1.91.0

aws-cdk.aws_apigateway>=1.91.0
aws-cdk.aws_cloudwatch>=1.91.0
aws-cdk.aws_iam>=1.91.0
aws-cdk.aws_kinesis>=1.91.0
aws-cdk.aws_kinesisanalytics>=1.91.0
aws-cdk.aws_kinesisfirehose>=1.91.0
aws-cdk.aws_s3>=1.91.0
aws-cdk.aws_glue>=1.91.0
aws-cdk.aws_athena>=1.91.0